+++ lib/src/brands/swift.dart
@@ -1,0 +1,16 @@
+import '../palette_rgb.dart';
+
+/// Swift's official brand colors
+///
+/// - Black;
+/// - Sunset Orange.
+///
+/// See also
+/// - [swift colors](https://brandpalettes.com/swift-colors/)
+class Swift extends PaletteRGB {
+  /// Sunset Orange #F05138.
+  const Swift.sunsetOrange() : super.opaque(0xF05138);
+
+  /// Black #000000.
+  const Swift.black() : super.black();
+}