+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - List of colors representing the brand's gradients.
 - The brands' official logos and icons downloaded from their official web sites.
 
+## [0.1.12] - 2021-04-19 [#29](https://github.com/rafamizes/flutter_brand_palettes/issues/29)
+### Added
+- Color palettes: FedEx, Messenger, Netflix, Reddit, Swift.
+- PaletteRGB class convenience constructors 'black' and 'white'.
+
 ### Changed
 - Less verbose and more concise changelog content.